requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py requirements-dev.in
#
click==6.7                # via pip-tools
fancycompleter==0.7       # via pdbpp
first==2.0.1              # via pip-tools
isort==4.2.5
jsonschema==2.6.0         # via target-csv
ordereddict==1.1          # via pdbpp
pdbpp==0.8.3
pip-tools==1.9.0
pipdeptree==0.10.1
pygments==2.2.0           # via pdbpp
singer-python==0.1.0
six==1.10.0               # via pip-tools
target-csv==0.2.2
wmctrl==0.3               # via pdbpp
